refactor: codebase cleanup — split views, remove legacy routes, consolidate tests
- Remove legacy email-level routes (8 URLs) and view functions (~600 lines)
  that duplicated the thread-based system
- Split views.py (3,104 lines) into views/ package:
  helpers.py, pages.py, threads.py, settings.py
- Extract _get_team_members() (19 sites) and _log_activity() (12 sites) helpers
- Consolidate 7 test files into their natural homes (45 → 38 files)
- Add test_state.py (14 tests for circuit breaker / StateManager)
- Guard _send_assignment_chat() and notify_mention() with
  chat_notifications_enabled check — prevents Chat spam during tests
- Fix test_disabled_when_threshold_not_set (delete seeded config first)
- Add deploy Chat notifications (deploying/deployed/failed) to deploy.yml
- Fix activity feed template to use thread URLs instead of removed email URLs

815 tests passing, 0 failures. Net -1,765 lines.
