--- apps/emails/tests/test_views.py
@@ -1,441 +1,0 @@
-"""Tests for the email dashboard views."""
-
-import pytest
-from datetime import datetime, timezone
-from django.test import Client
-from django.urls import reverse
-
-from apps.accounts.models import User
-from apps.emails.models import Email
-from conftest import create_email
-
-
-@pytest.fixture
-def admin_client(admin_user):
-    """Authenticated client logged in as admin."""
-    c = Client()
-    c.login(username="admin", password="testpass123")
-    return c
-
-
-@pytest.fixture
-def member_client(member_user):
-    """Authenticated client logged in as member."""
-    c = Client()
-    c.login(username="member", password="testpass123")
-    return c
-
-
-class TestEmailListAuth:
-    """Test authentication requirements for email_list view."""
-
-    def test_returns_200_for_authenticated_user(self, admin_client, db):
-        """Test 1: email_list view returns 200 for authenticated user."""
-        response = admin_client.get(reverse("emails:email_list"))
-        assert response.status_code == 200
-
-    def test_redirects_unauthenticated_user(self, client, db):
-        """Test 2: email_list view redirects to login for unauthenticated user."""
-        response = client.get(reverse("emails:email_list"))
-        assert response.status_code == 302
-        assert "/accounts/login/" in response.url
-
-
-class TestEmailListDefaultView:
-    """Test default view behavior based on user role."""
-
-    def test_admin_default_shows_unassigned(self, admin_client, admin_user, member_user, db):
-        """Test 3: Default admin view filters to unassigned emails."""
-        assigned = create_email(message_id="msg_assigned", assigned_to=member_user)
-        unassigned = create_email(message_id="msg_unassigned", assigned_to=None)
-
-        response = admin_client.get(reverse("emails:email_list"))
-        emails = list(response.context["emails"])
-        assert unassigned in emails
-        assert assigned not in emails
-
-    def test_member_default_shows_own(self, member_client, member_user, admin_user, db):
-        """Test 4: Default member view filters to own assigned emails."""
-        own = create_email(message_id="msg_own", assigned_to=member_user)
-        other = create_email(message_id="msg_other", assigned_to=admin_user)
-        unassigned = create_email(message_id="msg_none", assigned_to=None)
-
-        response = member_client.get(reverse("emails:email_list"))
-        emails = list(response.context["emails"])
-        assert own in emails
-        assert other not in emails
-        assert unassigned not in emails
-
-
-class TestEmailListFilters:
-    """Test filtering behavior."""
-
-    def test_filter_by_status(self, admin_client, db):
-        """Test 5: Filter by status returns only matching emails."""
-        new_email = create_email(message_id="msg_new", status=Email.Status.NEW)
-        closed_email = create_email(message_id="msg_closed", status=Email.Status.CLOSED)
-
-        response = admin_client.get(reverse("emails:email_list"), {"view": "all", "status": "new"})
-        emails = list(response.context["emails"])
-        assert new_email in emails
-        assert closed_email not in emails
-
-    def test_filter_by_priority(self, admin_client, db):
-        """Test 6: Filter by priority returns only matching emails."""
-        high = create_email(message_id="msg_high", priority="HIGH")
-        low = create_email(message_id="msg_low", priority="LOW")
-
-        response = admin_client.get(reverse("emails:email_list"), {"view": "all", "priority": "HIGH"})
-        emails = list(response.context["emails"])
-        assert high in emails
-        assert low not in emails
-
-    def test_combined_filters(self, admin_client, db):
-        """Test 11: Combined filter query params return only emails matching ALL criteria."""
-        match = create_email(
-            message_id="msg_match",
-            status=Email.Status.NEW,
-            priority="HIGH",
-            to_inbox="info@vidarbhainfotech.com",
-        )
-        wrong_status = create_email(
-            message_id="msg_wrong_status",
-            status=Email.Status.CLOSED,
-            priority="HIGH",
-            to_inbox="info@vidarbhainfotech.com",
-        )
-        wrong_priority = create_email(
-            message_id="msg_wrong_pri",
-            status=Email.Status.NEW,
-            priority="LOW",
-            to_inbox="info@vidarbhainfotech.com",
-        )
-        wrong_inbox = create_email(
-            message_id="msg_wrong_inbox",
-            status=Email.Status.NEW,
-            priority="HIGH",
-            to_inbox="sales@vidarbhainfotech.com",
-        )
-
-        response = admin_client.get(
-            reverse("emails:email_list"),
-            {"view": "all", "status": "new", "priority": "HIGH", "inbox": "info@vidarbhainfotech.com"},
-        )
-        emails = list(response.context["emails"])
-        assert match in emails
-        assert wrong_status not in emails
-        assert wrong_priority not in emails
-        assert wrong_inbox not in emails
-
-
-class TestEmailListSort:
-    """Test sorting behavior."""
-
-    def test_default_sort_by_created_at_desc(self, admin_client, db):
-        """Test 7a: Default sort is -created_at."""
-        old = create_email(
-            message_id="msg_old",
-            received_at=datetime(2026, 3, 1, tzinfo=timezone.utc),
-        )
-        new = create_email(
-            message_id="msg_new_sort",
-            received_at=datetime(2026, 3, 10, tzinfo=timezone.utc),
-        )
-
-        response = admin_client.get(reverse("emails:email_list"), {"view": "all"})
-        emails = list(response.context["emails"])
-        # Newer should come first (default sort is -created_at)
-        assert emails.index(new) < emails.index(old)
-
-    def test_sort_by_priority(self, admin_client, db):
-        """Test 7b: Sort by priority orders correctly."""
-        low = create_email(message_id="msg_low_sort", priority="LOW")
-        high = create_email(message_id="msg_high_sort", priority="HIGH")
-
-        response = admin_client.get(reverse("emails:email_list"), {"view": "all", "sort": "priority"})
-        emails = list(response.context["emails"])
-        assert len(emails) == 2
-
-
-class TestEmailListPagination:
-    """Test pagination behavior."""
-
-    def test_pagination_25_per_page(self, admin_client, db):
-        """Test 8: Pagination returns 25 per page."""
-        for i in range(30):
-            create_email(
-                message_id=f"msg_page_{i}",
-                received_at=datetime(2026, 3, 10, i % 24, 0, 0, tzinfo=timezone.utc),
-            )
-
-        response = admin_client.get(reverse("emails:email_list"), {"view": "all"})
-        assert len(response.context["emails"]) == 25
-
-    def test_filter_params_preserved_in_pagination(self, admin_client, db):
-        """Test 12: Filter state preserved in URL query params in pagination links."""
-        for i in range(30):
-            create_email(
-                message_id=f"msg_filter_page_{i}",
-                status=Email.Status.NEW,
-                priority="HIGH",
-            )
-
-        response = admin_client.get(
-            reverse("emails:email_list"),
-            {"view": "all", "status": "new", "priority": "HIGH"},
-        )
-        content = response.content.decode()
-        # Pagination link to page 2 should contain filter params
-        assert "status=new" in content
-        assert "priority=HIGH" in content
-
-
-class TestEmailListTabs:
-    """Test tab view behavior."""
-
-    def test_all_tab_shows_completed_emails(self, admin_client, db):
-        """Test 9: View tab 'all' shows all completed emails."""
-        completed = create_email(message_id="msg_completed")
-        pending = create_email(
-            message_id="msg_pending",
-            processing_status=Email.ProcessingStatus.PENDING,
-        )
-
-        response = admin_client.get(reverse("emails:email_list"), {"view": "all"})
-        emails = list(response.context["emails"])
-        assert completed in emails
-        assert pending not in emails
-
-
-class TestEmailListHTMX:
-    """Test HTMX partial response behavior."""
-
-    def test_htmx_returns_partial(self, admin_client, db):
-        """Test 10: HTMX request returns partial (no base layout), non-HTMX returns full page."""
-        create_email(message_id="msg_htmx")
-
-        # Non-HTMX: full page with sidebar
-        response = admin_client.get(reverse("emails:email_list"), {"view": "all"})
-        content = response.content.decode()
-        assert "VIPL Triage" in content  # Sidebar branding
-
-        # HTMX: partial without sidebar
-        response = admin_client.get(
-            reverse("emails:email_list"),
-            {"view": "all"},
-            HTTP_HX_REQUEST="true",
-        )
-        content = response.content.decode()
-        assert "VIPL Triage" not in content  # No sidebar in partial
-
-
-class TestEmailCountOOBUpdate:
-    """Test that HTMX responses include OOB email count update."""
-
-    def test_htmx_unassigned_view_has_oob_count(self, admin_client, db):
-        """HTMX request to unassigned view includes OOB count span."""
-        create_email(message_id="msg_oob_1")
-        response = admin_client.get(
-            reverse("emails:email_list"),
-            {"view": "unassigned"},
-            HTTP_HX_REQUEST="true",
-        )
-        content = response.content.decode()
-        assert 'id="email-count"' in content
-        assert 'hx-swap-oob' in content
-
-    def test_htmx_mine_view_has_oob_count(self, admin_client, admin_user, db):
-        """HTMX request to mine view includes OOB count span."""
-        create_email(message_id="msg_oob_2", assigned_to=admin_user)
-        response = admin_client.get(
-            reverse("emails:email_list"),
-            {"view": "mine"},
-            HTTP_HX_REQUEST="true",
-        )
-        content = response.content.decode()
-        assert 'id="email-count"' in content
-        assert 'hx-swap-oob' in content
-
-    def test_non_htmx_has_no_oob(self, admin_client, db):
-        """Non-HTMX full page render does NOT have hx-swap-oob on count."""
-        create_email(message_id="msg_oob_3")
-        response = admin_client.get(
-            reverse("emails:email_list"),
-            {"view": "all"},
-        )
-        content = response.content.decode()
-        assert 'hx-swap-oob' not in content
-
-    def test_oob_count_has_correct_pluralization(self, admin_client, db):
-        """OOB count shows correct plural form."""
-        create_email(message_id="msg_oob_single")
-        response = admin_client.get(
-            reverse("emails:email_list"),
-            {"view": "all"},
-            HTTP_HX_REQUEST="true",
-        )
-        content = response.content.decode()
-        assert "1 email" in content
-
-
-class TestWelcomeBanner:
-    """UX-01: Welcome banner with role-specific guidance."""
-
-    def test_banner_present_for_admin(self, admin_client, db):
-        """Welcome banner div is present in admin dashboard response."""
-        response = admin_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert 'id="welcome-banner"' in content
-
-    def test_banner_present_for_member(self, member_client, db):
-        """Welcome banner div is present in member dashboard response."""
-        response = member_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert 'id="welcome-banner"' in content
-
-    def test_admin_sees_assign_guidance(self, admin_client, db):
-        """Admin banner contains guidance about assigning emails."""
-        response = admin_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert "assign" in content.lower() or "Assign" in content
-
-    def test_member_sees_my_emails_guidance(self, member_client, db):
-        """Member banner contains guidance about My Emails."""
-        response = member_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert "My Emails" in content
-
-    def test_banner_has_session_storage_dismiss(self, admin_client, db):
-        """Banner has sessionStorage JS for dismiss logic."""
-        response = admin_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert "sessionStorage" in content
-        assert "vipl_welcome_dismissed" in content
-
-
-class TestFilterIndicators:
-    """UX-02: Active filter count and clear-all link."""
-
-    def test_one_filter_shows_indicator(self, admin_client, db):
-        """With one filter active, filter panel is open with Clear all link."""
-        create_email(message_id="msg_fi_1")
-        response = admin_client.get(
-            reverse("emails:email_list"),
-            {"view": "all", "status": "new"},
-        )
-        content = response.content.decode()
-        assert "Clear all" in content
-        assert 'id="filter-panel"' in content
-
-    def test_two_filters_show_indicator(self, admin_client, db):
-        """With two filters active, filter panel is open with Clear all link."""
-        create_email(message_id="msg_fi_2")
-        response = admin_client.get(
-            reverse("emails:email_list"),
-            {"view": "all", "status": "new", "priority": "HIGH"},
-        )
-        content = response.content.decode()
-        assert "Clear all" in content
-
-    def test_no_filters_no_indicator(self, admin_client, db):
-        """With no filters active, 'filter active' text is absent."""
-        create_email(message_id="msg_fi_3")
-        response = admin_client.get(
-            reverse("emails:email_list"),
-            {"view": "all"},
-        )
-        content = response.content.decode()
-        assert "filter active" not in content
-        assert "filters active" not in content
-
-    def test_clear_all_preserves_view(self, admin_client, db):
-        """Clear all link includes current view param."""
-        create_email(message_id="msg_fi_4")
-        response = admin_client.get(
-            reverse("emails:email_list"),
-            {"view": "all", "status": "new"},
-        )
-        content = response.content.decode()
-        assert "Clear all" in content
-        assert "?view=all" in content
-
-
-class TestKeyboardNav:
-    """UX-04: Keyboard navigation between email cards."""
-
-    def test_response_contains_arrow_down_handler(self, admin_client, db):
-        """Response JS contains ArrowDown key handling."""
-        response = admin_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert "ArrowDown" in content
-
-    def test_response_contains_arrow_up_handler(self, admin_client, db):
-        """Response JS contains ArrowUp key handling."""
-        response = admin_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert "ArrowUp" in content
-
-    def test_response_contains_escape_close_detail(self, admin_client, db):
-        """Response JS contains Escape key handler referencing closeDetail."""
-        response = admin_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert "Escape" in content
-        assert "closeDetail" in content
-
-    def test_response_contains_form_field_guard(self, admin_client, db):
-        """Response JS guards against keyboard nav when in form fields."""
-        response = admin_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert "INPUT" in content
-        assert "activeElement" in content
-
-    def test_response_contains_role_article_selector(self, admin_client, db):
-        """Response JS uses role=article selector for email cards."""
-        response = admin_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert 'role="article"' in content or "role=\\\"article\\\"" in content
-
-
-class TestLoadingSkeleton:
-    """UX-05: Loading skeleton for detail panel."""
-
-    def test_response_contains_before_request_listener(self, admin_client, db):
-        """Response JS has htmx:beforeRequest event listener."""
-        response = admin_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert "htmx:beforeRequest" in content
-
-    def test_skeleton_targets_detail_panel(self, admin_client, db):
-        """Skeleton injection logic targets detail-panel specifically."""
-        response = admin_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert "detail-panel" in content
-
-    def test_skeleton_has_animate_pulse(self, admin_client, db):
-        """Skeleton HTML contains animate-pulse class."""
-        response = admin_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert "animate-pulse" in content
-
-    def test_skeleton_has_slate_placeholder_blocks(self, admin_client, db):
-        """Skeleton HTML uses bg-slate-200 placeholder blocks."""
-        response = admin_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert "bg-slate-200" in content
-
-
-class TestScrollSnap:
-    """UX-03: Scroll-snap on stat cards for mobile."""
-
-    def test_stat_container_has_snap_classes(self, admin_client, db):
-        """Stat cards container has snap-x and snap-mandatory classes."""
-        response = admin_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert "snap-x" in content
-        assert "snap-mandatory" in content
-
-    def test_stat_cards_have_snap_start(self, admin_client, db):
-        """Each stat card has snap-start class."""
-        response = admin_client.get(reverse("emails:email_list"))
-        content = response.content.decode()
-        assert "snap-start" in content
