--- templates/emails/_assign_dropdown.html
@@ -1,25 +1,0 @@
-<!-- Inline assignment dropdown for email card (admin only) -->
-<form hx-post="{% url 'emails:assign_email' email.pk %}"
-      hx-target="#card-{{ email.pk }}"
-      hx-swap="outerHTML"
-      class="inline"
-      onclick="event.stopPropagation();">
-    {% csrf_token %}
-    <select name="assignee_id"
-            onchange="this.form.requestSubmit();"
-            class="vipl-select text-[10px] font-semibold rounded-lg px-2 py-1 cursor-pointer hover:border-[var(--vipl-primary)] transition-all max-w-[140px]">
-        {% if email.assigned_to %}
-        <option value="{{ email.assigned_to.pk }}" selected>{{ email.assigned_to.get_full_name|default:email.assigned_to.username }}</option>
-        <option value="" class="text-[var(--vipl-text-dim)]">-- Unassign --</option>
-        {% else %}
-        <option value="">Assign to...</option>
-        {% endif %}
-        {% for member in team_members %}
-            {% if member.pk != email.assigned_to_id %}
-            <option value="{{ member.pk }}">
-                {{ member.get_full_name|default:member.username }}{% if member.role == "admin" %} (Admin){% endif %}
-            </option>
-            {% endif %}
-        {% endfor %}
-    </select>
-</form>
