--- templates/emails/_email_card.html
@@ -1,97 +1,0 @@
-{% load email_tags %}
-
-<div id="card-{{ email.pk }}"
-     {% if oob %}hx-swap-oob="outerHTML"{% endif %}
-     class="card-hover border-l-[3px] {{ email.priority|priority_border }} bg-white mx-2 my-1 rounded-lg px-3 py-2.5 cursor-pointer border border-slate-200/40 hover:border-slate-200"
-     hx-get="{% url 'emails:email_detail' email.pk %}"
-     hx-target="#detail-panel"
-     hx-swap="innerHTML"
-     role="article" tabindex="0" aria-label="{{ email.subject }}"
-     onkeydown="if(event.key==='Enter'||event.key===' '){event.preventDefault();this.click();}"
-     onclick="document.querySelectorAll('.card-selected').forEach(c=>c.classList.remove('card-selected'));this.classList.add('card-selected');">
-
-    <!-- Row 1: sender, category, time -->
-    <div class="flex items-center justify-between gap-2 mb-1">
-        <div class="flex items-center gap-2 min-w-0">
-            <div class="w-5.5 h-5.5 rounded-full bg-gradient-to-br from-{{ email.priority|priority_base }}-400 to-{{ email.priority|priority_base }}-500 flex items-center justify-center text-white text-[8px] font-bold shrink-0" style="width:22px;height:22px;">
-                {{ email.from_name.0|default:"?" }}
-            </div>
-            <div class="min-w-0 truncate">
-                <span class="text-[12px] font-semibold text-slate-800 truncate block">{{ email.from_name|default:email.from_address }}</span>
-                {% if email.from_name and email.from_address %}
-                <span class="text-[10px] text-slate-400 truncate block">{{ email.from_address }}</span>
-                {% endif %}
-            </div>
-        </div>
-        <div class="flex items-center gap-1.5 shrink-0">
-            {% if email.category %}
-            <span class="px-1.5 py-0.5 rounded text-[8px] font-bold bg-slate-100 text-slate-500 uppercase tracking-wide">{{ email.category }}</span>
-            {% endif %}
-            <span class="text-[10px] text-slate-400 tabular-nums font-medium">{{ email.created_at|time_ago }}</span>
-        </div>
-    </div>
-
-    <!-- Row 2: subject -->
-    <div class="text-[12px] text-slate-800 truncate font-semibold leading-snug mb-0.5 pl-[30px]">{{ email.subject }}</div>
-
-    <!-- Row 3: AI summary (compact) -->
-    {% if email.ai_summary %}
-    <div class="text-[11px] text-slate-400 truncate leading-relaxed mb-1.5 pl-[30px]">{{ email.ai_summary }}</div>
-    {% else %}
-    <div class="mb-1"></div>
-    {% endif %}
-
-    <!-- Row 4: status badges + SLA + assignee -->
-    <div class="flex items-center justify-between pl-[30px]">
-        <div class="flex items-center gap-1">
-            <span class="inline-flex items-center px-1.5 py-0.5 rounded text-[8px] font-bold uppercase tracking-wider bg-{{ email.priority|priority_base }}-50 text-{{ email.priority|priority_base }}-700" title="{{ email.priority|priority_tooltip }}">
-                {{ email.priority|default:"--" }}
-            </span>
-            <span class="inline-flex items-center gap-1 px-1.5 py-0.5 rounded text-[8px] font-semibold bg-{{ email.status|status_base }}-50 text-{{ email.status|status_base }}-600" title="{{ email.status|status_tooltip }}">
-                <span class="w-1 h-1 rounded-full bg-{{ email.status|status_base }}-500"></span>
-                {{ email.get_status_display }}
-            </span>
-            {% if email.is_spam %}
-            <span class="inline-flex items-center px-1.5 py-0.5 rounded text-[8px] font-bold uppercase tracking-wider bg-red-100 text-red-700" title="Flagged by spam filter — review and whitelist if legitimate">Spam</span>
-            {% endif %}
-            {% if email.sla_respond_deadline %}
-            <span class="inline-flex items-center gap-0.5 px-1.5 py-0.5 rounded text-[8px] font-bold text-{{ email.sla_respond_deadline|sla_color }}-700 bg-{{ email.sla_respond_deadline|sla_color }}-50">
-                <svg class="w-2.5 h-2.5" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2.5" d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z"/></svg>
-                {{ email.sla_respond_deadline|sla_countdown }}
-            </span>
-            {% endif %}
-            {% if email.ai_suggested_assignee.name and not email.assigned_to %}
-            <span class="px-1.5 py-0.5 rounded text-[8px] font-semibold bg-blue-50 text-blue-600">
-                AI: {{ email.ai_suggested_assignee.name }}
-            </span>
-            {% endif %}
-        </div>
-
-        <div class="shrink-0 flex items-center gap-1" onclick="event.stopPropagation();">
-            {% if not email.assigned_to and user.can_assign or not email.assigned_to and email.category in user_visible_categories %}
-            <form hx-post="{% url 'emails:claim_email' email.pk %}"
-                  hx-target="#card-{{ email.pk }}"
-                  hx-swap="outerHTML"
-                  class="inline">
-                {% csrf_token %}
-                <button type="submit" hx-disabled-elt="this" class="px-2 py-0.5 text-[9px] font-bold text-emerald-700 bg-emerald-50 rounded border border-emerald-200/60 hover:bg-emerald-100 transition-colors cursor-pointer">
-                    Claim
-                </button>
-            </form>
-            {% endif %}
-
-            {% if user.can_assign %}
-                {% include "emails/_assign_dropdown.html" with email=email team_members=team_members %}
-            {% elif email.assigned_to %}
-            <div class="flex items-center gap-1">
-                <div class="w-4 h-4 rounded-full bg-gradient-to-br from-primary-400 to-primary-500 flex items-center justify-center text-white text-[7px] font-bold" style="width:16px;height:16px;">
-                    {{ email.assigned_to.first_name.0|default:email.assigned_to.username.0|upper }}
-                </div>
-                <span class="text-[10px] font-medium text-slate-500">{{ email.assigned_to.first_name|default:email.assigned_to.username }}</span>
-            </div>
-            {% else %}
-            <span class="text-[9px] font-bold text-amber-600 bg-amber-50 px-1.5 py-0.5 rounded">Unassigned</span>
-            {% endif %}
-        </div>
-    </div>
-</div>
